+++ capycli/dependencies/python.py
@@ -1,5 +1,5 @@
 # -------------------------------------------------------------------------------
-# Copyright (c) 2019-2025 Siemens
+# Copyright (c) 2019-2026 Siemens
 # All Rights Reserved.
 # Author: thomas.graf@siemens.com
 #
@@ -19,10 +19,13 @@ from typing import Any, Dict, List, Optional
 import chardet
 import requests
 import requirements
-from cyclonedx.factory.license import LicenseFactory
-from cyclonedx.model import ExternalReference, ExternalReferenceType, HashType, Property, XsUri
+from cyclonedx.contrib.hash.factories import HashTypeFactory
+from cyclonedx.contrib.license.factories import LicenseFactory
+from cyclonedx.model import ExternalReference, ExternalReferenceType, Property, XsUri
 from cyclonedx.model.bom import Bom
 from cyclonedx.model.component import Component
+from cyclonedx.model.contact import OrganizationalEntity
+from cyclonedx.model.lifecycle import LifecyclePhase, PredefinedLifecycle
 from halo import Halo
 from packageurl import PackageURL
 
@@ -310,6 +313,18 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
                 cxcomp.licenses.add(license_factory.make_with_name(license))
                 LOG.debug("  got license")
 
+            author = meta["info"].get("author", "")
+            cxcomp.supplier = OrganizationalEntity()
+            if author:
+                cxcomp.author = author
+                LOG.debug("  got author")
+
+                # set also also as supplier
+                cxcomp.supplier.name = author
+            else:
+                cxcomp.author = "N/A"
+                cxcomp.supplier.name = "N/A"
+
             data = meta["info"].get("summary", "")
             if data:
                 cxcomp.description = data
@@ -355,6 +370,8 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
                         cxcomp.external_references.add(ext_ref)
                         LOG.debug("  got GitHub source file url")
 
+            hf = HashTypeFactory()
+            component_hashes = []
             if "urls" in meta:
                 # there can be multiple entries, for wheel, source, etc.
                 for item in meta["urls"]:
@@ -363,17 +380,25 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
                             file_uri = item["filename"]
                             if not file_uri.startswith("file://"):
                                 file_uri = "file:///" + file_uri
+                            hashes = []
+                            if "digests" in item and "sha256" in item["digests"] and item["digests"]["sha256"]:
+                                hashes = [hf.from_hashlib_alg("sha256", item["digests"]["sha256"])]
                             ext_ref = ExternalReference(
                                 type=ExternalReferenceType.DISTRIBUTION,
                                 comment=CaPyCliBom.BINARY_FILE_COMMENT,
-                                url=XsUri(file_uri))
+                                url=XsUri(file_uri),
+                                hashes=hashes)
                             cxcomp.external_references.add(ext_ref)
                             LOG.debug("  got binary file")
+
+                            # take the wheel hash also as component hash
+                            component_hashes = hashes
 
                             ext_ref = ExternalReference(
                                 type=ExternalReferenceType.DISTRIBUTION,
                                 comment=CaPyCliBom.BINARY_URL_COMMENT,
-                                url=XsUri(item["url"]))
+                                url=XsUri(item["url"]),
+                                hashes=hashes)
                             cxcomp.external_references.add(ext_ref)
                             LOG.debug("  got binary file url")
 
@@ -381,10 +406,14 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
                             file_uri = item["filename"]
                             if not file_uri.startswith("file://"):
                                 file_uri = "file:///" + file_uri
+                            hashes = []
+                            if "digests" in item and "sha256" in item["digests"] and item["digests"]["sha256"]:
+                                hashes = [hf.from_hashlib_alg("sha256", item["digests"]["sha256"])]
                             ext_ref = ExternalReference(
                                 type=ExternalReferenceType.DISTRIBUTION,
                                 comment=CaPyCliBom.SOURCE_FILE_COMMENT,
-                                url=XsUri(file_uri))
+                                url=XsUri(file_uri),
+                                hashes=hashes)
                             cxcomp.external_references.add(ext_ref)
                             LOG.debug("  got source file")
 
@@ -392,9 +421,19 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
                                 ext_ref = ExternalReference(
                                     type=ExternalReferenceType.DISTRIBUTION,
                                     comment=CaPyCliBom.SOURCE_URL_COMMENT,
-                                    url=XsUri(item["url"]))
+                                    url=XsUri(item["url"]),
+                                    hashes=hashes)
                                 cxcomp.external_references.add(ext_ref)
                                 LOG.debug("  got sdist source file url")
+
+            if component_hashes:
+                # add the hashes also to the component itself
+                for h in component_hashes:
+                    cxcomp.hashes.add(h)
+
+            # There is no copyright information in Python package metadata.
+            # There is also no copyright information available from GitHub
+            cxcomp.copyright = "N/A"
 
     def convert_package_list(self, package_list: List[Dict[str, Any]], search_meta_data: bool,
                              package_source: str = "") -> Bom:
@@ -443,7 +482,7 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
         for package in package_list:
             print_text("  " + package["name"] + ", " + package["version"])
 
-        print()
+        print_text()
 
     def determine_file_type(self, full_filename: str) -> InputFileType:
         """
@@ -697,12 +736,13 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
                 for file_metadata in package.files:
                     LOG.debug(f"    Processing file_metadata: {file_metadata}")
                     try:
+                        hf = HashTypeFactory()
                         cxcomp.external_references.add(ExternalReference(
                             type=ExternalReferenceType.DISTRIBUTION,
                             url=XsUri(cxcomp.get_pypi_url()),
                             # comment=f'Distribution file: {file_metadata.file}',
                             comment=CaPyCliBom.BINARY_URL_COMMENT,
-                            hashes=[HashType.from_composite_str(file_metadata.hash)]
+                            hashes=[hf.from_composite_str(file_metadata.hash)]
                         ))
                     except Exception as ex:
                         # IGNORE
@@ -759,12 +799,13 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
                 for file_metadata in package.files:
                     LOG.debug(f"    Processing file_metadata: {file_metadata}")
                     try:
+                        hf = HashTypeFactory()
                         cxcomp.external_references.add(ExternalReference(
                             type=ExternalReferenceType.DISTRIBUTION,
                             url=XsUri(cxcomp.get_pypi_url()),
                             # comment=f'Distribution file: {file_metadata.file}',
                             comment=CaPyCliBom.BINARY_URL_COMMENT,
-                            hashes=[HashType.from_composite_str(file_metadata.hash)]
+                            hashes=[hf.from_composite_str(file_metadata.hash)]
                         ))
                     except Exception as ex:
                         # IGNORE
@@ -778,6 +819,103 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
             spinner.stop()
 
         return sbom
+
+    def add_main_component(self, filename: str, sbom: Bom) -> None:
+        """Add main project component to the SBOM."""
+        folder = os.path.dirname(filename)
+
+        if self.proj_file_override:
+            # override for unit tests
+            pyproject_file = self.proj_file_override
+        else:
+            pyproject_file = os.path.join(folder, "pyproject.toml")
+        pyproject_info = self.read_pyproject_file(pyproject_file)
+        if not pyproject_info:
+            return
+
+        new_pyproject_format = False
+        if "project" in pyproject_info:
+            # this is for poetry >= 2.0 and uv
+            cfg = pyproject_info["project"]
+            new_pyproject_format = True
+        else:
+            cfg = pyproject_info["tool"]["poetry"]
+
+        name = cfg.get("name", "").strip()
+        version = cfg.get("version", "").strip()
+        if not name or not version:
+            LOG.warning("Main project component has no name or version. Skipping.")
+            return
+
+        # special hack for CaPyCLI
+        copyright = "N/A"
+        if name == "capycli":
+            copyright = "Siemens AG"
+
+        description = cfg.get("description", "").strip()
+        purl = self.generate_purl(name, version)
+        cxcomp = Component(
+            name=name,
+            version=version,
+            purl=PackageURL.from_string(purl),
+            copyright=copyright,
+            bom_ref=purl,
+            description=description)
+        sbom.metadata.component = cxcomp
+        for c in sbom.components:
+            sbom.register_dependency(sbom.metadata.component, [c])
+
+        license = cfg.get("license", "").strip()
+        if license:
+            license_factory = LicenseFactory()
+            cxcomp.licenses.add(license_factory.make_with_name(license))
+            LOG.debug("  got license")
+
+        authors = cfg.get("authors", [])
+        author = ""
+        if new_pyproject_format:
+            if authors and len(authors) > 0:
+                author = authors[0].get("name", "") if isinstance(authors[0], dict) else authors[0].strip()
+        else:
+            if authors and len(authors) > 0:
+                author = authors[0].strip()
+
+        cxcomp.supplier = OrganizationalEntity()
+        if author:
+            cxcomp.author = author
+            LOG.debug("  got author")
+
+            # set also also as supplier
+            cxcomp.supplier.name = author
+        else:
+            cxcomp.author = "N/A"
+            cxcomp.supplier.name = "N/A"
+
+        # authors = [
+        #    {name = "Thomas Graf", email="thomas.graf@siemens.com"},
+        #    {name = "Gernot Hillier", email="gernot.hillier@siemens.com"}
+        # ]
+        # [project.urls]
+        # repository = "https://github.com/sw360/capycli"
+        # homepage = "https://github.com/sw360/capycli"
+        # issues = "https://github.com/sw360/capycli/issues"
+
+    def do_json_post_processing(self, filename: str) -> None:
+        """Do some JSON post-processing to add composition information."""
+        LOG.debug("Doing JSON post-processing to add composition information")
+        try:
+            json_sbom: Dict[str, Any] = capycli.common.json_support.load_json_file(filename)
+
+            # add composition information
+            if "components" in json_sbom and len(json_sbom["components"]) > 0:
+                json_sbom["compositions"] = [
+                    {
+                        "aggregate": "unknown"
+                    }
+                ]
+            capycli.common.json_support.write_json_to_file(json_sbom, filename)
+        except Exception as ex:
+            print_red("Error postprocessing SBOM file: " + repr(ex))
 
     @staticmethod
     def check_meta_data(sbom: Bom, verbose: bool) -> bool:
@@ -842,18 +980,18 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
             " - Determine Python components/dependencies\n")
 
         if args.help:
-            print("usage: capycli getdependencies python [-i INPUTFILE] [-o OUTFILE] [-ol] [-v] [-ds]")
-            print("")
-            print("Determine Python project dependencies")
-            print("")
-            print("optional arguments:")
-            print("    -h, --help            show this help message and exit")
-            print("    -i INPUTFILE, --inputfile INPUTFILE")
-            print("                            input file (requirements.txt)")
-            print("    -o OUTFILE, --outfile OUTFILE")
-            print("                            output file (BOM)")
-            print("    -v, --verbose         verbose output")
-            print("    --search-meta-data    search for package meta data")
+            print_text("usage: capycli getdependencies python [-i INPUTFILE] [-o OUTFILE] [-ol] [-v] [-ds]")
+            print_text("")
+            print_text("Determine Python project dependencies")
+            print_text("")
+            print_text("optional arguments:")
+            print_text("    -h, --help            show this help message and exit")
+            print_text("    -i INPUTFILE, --inputfile INPUTFILE")
+            print_text("                            input file (requirements.txt)")
+            print_text("    -o OUTFILE, --outfile OUTFILE")
+            print_text("                            output file (BOM)")
+            print_text("    -v, --verbose         verbose output")
+            print_text("    --search-meta-data    search for package meta data")
             return
 
         if not args.inputfile:
@@ -885,13 +1023,26 @@ class GetPythonDependencies(capycli.common.script_base.ScriptBase):
             print_text("Formatting package list...")
             sbom = self.convert_package_list(package_list, args.search_meta_data, args.package_source)
 
+        # add meta data for the main project component
+        if (datatype == InputFileType.POETRY_LOCK) or (datatype == InputFileType.UV_LOCK):
+            self.add_main_component(args.inputfile, sbom)
+
+        # add some more extras
+        # author of a created SBOM is always CaPyCLI
+        sbom.metadata.authors.add(OrganizationalEntity(name="CaPyCLI"))
+        # lifecycle phase is "build"
+        sbom.metadata.lifecycles.add(PredefinedLifecycle(phase=LifecyclePhase.BUILD))
+        # compositions are not supported by cyclonedx-pathon-lib
+        # We do this as a JSON post-processing step
+
         GetPythonDependencies.check_meta_data(sbom, self.verbose)
 
         if self.verbose:
-            print()
+            print_text()
 
         print_text("Writing new SBOM to " + args.outputfile)
         SbomWriter.write_to_json(sbom, args.outputfile, True)
+        self.do_json_post_processing(args.outputfile)
         print_text(" " + self.get_comp_count_text(sbom) + " items written to file.")
 
-        print()
+        print_text()